app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.2
djangorestframework==3.12.4
gunicorn==20.0.4
idna==3.4
pip==23.2.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
